Only delete glossary entries if glossary exist

+++ gp-includes/things/translation-set.php
@@ -755,10 +755,13 @@ class GP_Translation_Set extends GP_Thing {
 
 		// Get the single Project Glossary for the translation Set.
 		$glossaries = GP::$glossary->find( array( 'translation_set_id' => $this->id ) );
-		$glossary   = $glossaries[0];
 
-		// Delete the Project Glossary entries.
-		GP::$glossary_entry->delete_many( array( 'glossary_id' => $glossary->id ) );
+		if ( is_array( $glossaries ) && ! empty( $glossaries ) ) {
+			$glossary = $glossaries[0];
+
+			// Delete the Project Glossary entries.
+			GP::$glossary_entry->delete_many( array( 'glossary_id' => $glossary->id ) );
+		}
 
 		// Delete Translation Set Glossary.
 		GP::$glossary->delete_many( array( 'translation_set_id' => $this->id ) );
